Add custom font support, add example app
